change weight from int to double

+++ dg-muscle-ios/dg-muscle-ios/model/ExerciseHistory.swift
@@ -18,9 +18,9 @@ struct ExerciseHistory: Codable, Identifiable, Equatable {
     let records: [Record]
     let createdAt: CreatedAt?
     
-    var volume: Int {
+    var volume: Double {
         let allSets: [ExerciseSet] = records.map({ $0.sets }).flatMap({ $0 })
-        return allSets.reduce(0, { $0 + ($1.reps * $1.weight) })
+        return allSets.reduce(0, { $0 + (Double($1.reps) * $1.weight) })
     }
     
     var dateValue: Date? {
